fix loadSetting + read camera device id by arg

+++ include/projector_tracker/ProjectorTracker.h
@@ -88,7 +88,7 @@ private:
     int arucoH;
     int patternW;
     int patternH;
-    int squareSize;
+    double squareSize;
 };
 
 #endif // PROJECTORTRACKER_H
